Factor impossible games in weighting
Certain types of games were being underrepresented in the case where many decks didn't have enough cards of interest. For example if you want to test decks with 10 cards of interest and most decks only have 5-6 cards of interest games where 7+ cards of interest are drawn would be underrepresented in this sample. This change adds a new factor to the game weight to take this into account.